Added new methods to StateFactory
Create Primitive or Conservative states based on an input array that contains state data of type array_state_type, which is useful when wanting to create State objects from numpy arrays that hold state data in a certain variable basis. These functions are then used to create the left and right states for the riemann solver, which now only accepts PrimitiveState objects as state inputs.